NSQ Kumar2024: harden FleetAuthority evidence ingress
