+++ .github/workflows/nsq-kumar2024-promoted-binding.yml
@@ -16,7 +16,7 @@ on:
       - "packages/neuros/src/neuros/evidence/kumar2024_materialized_study.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_comparison.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_promoted_execution.py"
-      - "packages/neuros/src/neuros/evidence/kumar2024_promoted_fleet.py"
+      - "packages/neuros/src/neuros/evidence/*kumar2024_promoted_fleet*.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_promoted_binding.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_promoted_worker.py"
       - ".github/workflows/nsq-kumar2024-promoted-binding.yml"
@@ -36,7 +36,7 @@ on:
       - "packages/neuros/src/neuros/evidence/kumar2024_materialized_study.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_comparison.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_promoted_execution.py"
-      - "packages/neuros/src/neuros/evidence/kumar2024_promoted_fleet.py"
+      - "packages/neuros/src/neuros/evidence/*kumar2024_promoted_fleet*.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_promoted_binding.py"
       - "packages/neuros/src/neuros/evidence/kumar2024_promoted_worker.py"
       - ".github/workflows/nsq-kumar2024-promoted-binding.yml"
